First pass at windows fix for issue 27.
Broken, disabling buildbot, checking in so I can go to an appointment and work from elsewhere afterwards. More information on future checkin, refer to bug report.

+++ Core/ConnectionSML/src/sock_ClientSocket.h
@@ -32,7 +32,7 @@ public:
 	// Description	  : Connect to a server
 	//
 	/////////////////////////////////////////////////////////////////////
-	bool	ConnectToServer(char const* netAddress, unsigned short port) ;
+	bool	ConnectToServer(char const* netAddress, int port) ;
 };
 
 } // Namespace
